Moves inspection algorithms into evaluators.
