+++ src/IsAwaitable/IsKnownAwaitableExtension.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Threading.Tasks;
 
 namespace IsAwaitable
 {
@@ -43,11 +42,8 @@ namespace IsAwaitable
         /// </exception>
         public static bool IsKnownAwaitable(this Type type)
         {
-            var evaluation = GetEvaluationFor(type);
-            
-            return
-                evaluation == TypeEvaluation.Awaitable ||
-                evaluation == TypeEvaluation.AwaitableWithResult;
+            var evaluation = KnownAwaitableEvaluator.Evaluate(type);
+            return evaluation.IsAwaitable;
         }
 
         /// <summary>
@@ -82,52 +78,8 @@ namespace IsAwaitable
         /// </exception>
         public static bool IsKnownAwaitableWithResult(this Type type)
         {
-            var evaluation = GetEvaluationFor(type);
-            return evaluation == TypeEvaluation.AwaitableWithResult;
-        }
-
-        internal static TypeEvaluation GetEvaluationFor(this Type type)
-        {
-            if (type is null)
-                throw new ArgumentNullException(nameof(type));
-
-            // Known awaitable types are: Task, Task<>, ValueTask, ValueTask<>.
-
-            if (type.IsGenericType)
-            {
-                if (type.IsSubclassOfRawGeneric(typeof(Task<>)))
-                    return TypeEvaluation.AwaitableWithResult;
-
-                var genericTypeDefinition = type.GetGenericTypeDefinition();
-
-                if (typeof(ValueTask<>) == genericTypeDefinition)
-                    return TypeEvaluation.AwaitableWithResult;
-            }
-
-            if (typeof(Task).IsAssignableFrom(type) ||
-                typeof(ValueTask) == type)
-                return TypeEvaluation.Awaitable;
-
-            return TypeEvaluation.NotAwaitable;
-        }
-
-        private static bool IsSubclassOfRawGeneric(this Type type, Type rawGenericType)
-        {
-            var toCheck = type;
-
-            while (toCheck is { } && toCheck != typeof(object))
-            {
-                var target = toCheck.IsGenericType
-                    ? toCheck.GetGenericTypeDefinition()
-                    : toCheck;
-                
-                if (rawGenericType == target)
-                    return true;
-
-                toCheck = toCheck.BaseType;
-            }
-
-            return false;
+            var evaluation = KnownAwaitableEvaluator.Evaluate(type);
+            return evaluation.IsAwaitableWithResult;
         }
     }
 }
